Hotfix / Reference the right style partial
The wrongly named file was breaking the docker build.
Change the old name of the partial that styles a panel in a list of panels.
Previously it was named to be refering to the blog landing page but then was renamed to have a more general name in case it could be re-used. And it wasn’t updated on the styles loading file.
Oppsie