Rename guide to section ARIA on side nav
Changed the side nav contents list ARIA to be ’Pages in this section’, added aria-lebbeled by with the H2 reference

Made the heading visible and removed the top heading duplicate.

Now the structure follows the guidance:
https://www.w3.org/TR/wai-aria-practices/examples/landmarks/navigation.html

Before:
The ‘guide’ name refers to the example from ‘Mini Hub’ nhs.uk design pattern.
The ‘contents’ component needs an ARIA name that is more contextual.

Left ‘as is’ the ‘cms/core/templates/blocks/jump_menu_block.html’ as I’m not sure how and where it is used.