WIP: Document production release process
https://trello.com/c/1RWmUJeU/238-set-up-deployment-pipeline-xl

Currently the `main` branch of this repo is deployed to staging via Github
Actions.

Eventually the application will need a production release process.

Normally, the release process is divided as such:
- staging reflects the `develop` branch
- production reflects the `main` branch

And therefore `develop` will be automatically deployed to staging, not `main`.

Assuming in future this repo will have a `develop` branch created, and the
`develop` branch is deployed to staging, and the `main` branch is deployed to
production, this PR can be reviewed and merged.